Updated budget tracker
I made a few changes to Budget Tracker. First, I changed the variable names. The current variable names are not very descriptive. Instead of 'keep_going', I used 'continue_input', and instead of 'total_expense', I used 'sum_of_expenses.' Second, I noticed the variables'difference' and 'expensetotal' are not used in the code and I removed them. Third, I used f strings to format output instead of concatenating strings with varibles using the '+' operator. By using f-strings to format the output it makes the code easier to read and maintain. Lastly, instead of using a while loop with a condition that changes inside the loop, I use while true loop and break out of it when the user is done entering expenses.